finished Users API

+++ EasyGradeManager/Controllers/API/UsersController.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Data.Entity;
+﻿using EasyGradeManager.Models;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Web.Http;
-using EasyGradeManager.Models;
 using static EasyGradeManager.Static.Authorize;
+using static System.Data.Entity.EntityState;
 
 namespace EasyGradeManager.Controllers.API
 {
@@ -40,17 +40,17 @@ namespace EasyGradeManager.Controllers.API
             if (authorizedUser == null || authorizedUser.Id != id)
                 return Unauthorized();
             User user = db.Users.Find(id);
-            if (userDTO == null || user == null || !ModelState.IsValid || id != userDTO.Id)
+            if (userDTO == null || user == null || !ModelState.IsValid || id != userDTO.Id || !userDTO.Validate(false))
                 return BadRequest(ModelState);
             bool logoutNecessary = userDTO.Update(user);
-            string error = db.Update(user);
+            string error = db.Update(user, Modified);
             if (error != null)
                 return BadRequest(error);
             if (userDTO.NewRole != null)
             {
                 user = db.Users.Find(id);
                 userDTO.UpdateRole(user, db);
-                error = db.Update(user);
+                error = db.Update(user, Modified);
                 if (error != null)
                     return BadRequest(error);
             }
@@ -63,16 +63,25 @@ namespace EasyGradeManager.Controllers.API
         public IHttpActionResult PostUser(UserDetailDTO userDTO)
         {
             User authorizedUser = GetAuthorizedUser(Request.Headers.GetCookies("user").FirstOrDefault());
-            if (authorizedUser == null || !ModelState.IsValid || userDTO.Identifier.Contains("&") || userDTO.NewPassword.Contains("&") || 
-                userDTO.NewRole == null || (userDTO.NewRole != "Teacher" && userDTO.NewRole != "Tutor" && userDTO.NewRole != "Student"))
-                return BadRequest(ModelState);
-            User user = new User();
-            userDTO.Update(user);
+            if (authorizedUser == null)
+                return Unauthorized();
+            if (!ModelState.IsValid || !userDTO.Validate(true))
+                return BadRequest();
+            User user = userDTO.Create();
             db.Users.Add(user);
-            db.SaveChanges();
+            string error = db.Update(user, Added);
+            if (error != null)
+                return BadRequest(error);
             userDTO.UpdateRole(user, db);
-            db.Entry(user).State = EntityState.Modified;
-            db.SaveChanges();
+            error = db.Update(user, Modified);
+            if (error != null)
+            {
+                db.Users.Remove(user);
+                string error2 = db.Update(user, Deleted);
+                if (error2 != null)
+                    return BadRequest(error + '\n' + error2);
+                return BadRequest(error);
+            }
             UserListDTO result = new UserListDTO(user);
             return CreatedAtRoute("DefaultApi", new { id = userDTO.Id }, result);
         }
@@ -80,18 +89,21 @@ namespace EasyGradeManager.Controllers.API
         public IHttpActionResult DeleteUser(int id)
         {
             User authorizedUser = GetAuthorizedUser(Request.Headers.GetCookies("user").FirstOrDefault());
-            //if (authorizedUser == null || authorizedUser.Id != id)
-            //    return Unauthorized();
+            if (authorizedUser == null || authorizedUser.Id != id)
+                return Unauthorized();
             User user = db.Users.Find(id);
             if (user == null)
-            {
                 return NotFound();
-            }
-
+            if (!user.DeleteRoles(db))
+                return BadRequest();
+            string error = db.Update(user, Modified);
+            if (error != null)
+                return BadRequest(error);
             db.Users.Remove(user);
-            db.SaveChanges();
-
-            return Ok(user);
+            error = db.Update(user, Deleted);
+            if (error != null)
+                return BadRequest(error);
+            return Redirect("https://" + Request.RequestUri.Host + ":" + Request.RequestUri.Port + "/Logout");
         }
 
         protected override void Dispose(bool disposing)
